requirements.txt
fastapi
uvicorn[standard]
python-binance
aiohttp
pydantic
requests
